Implemented supervised autonomy; added component trigger
The manager is now only active if the component is triggered; this also
triggers components that are required for state estimation.

Additionally, the component can be used in either fully autonomous mode
(in which case the robot's proposed action is always executed) or in a
supervised autonomous mode (where the proposed action has to be approved
by a therapist).