[game_library/game_base] Add ack msg latching; prevent processing status/answer msgs multiple times
Latching was added in order to prevent messages from being lost on the
subscriber side. Without the latching, this could happen even with the
resending of messages - the publisher and subscriber would sometimes be
completely out of sync, so the resending wouldn't help; latching avoids
this problem.

To prevent answer and status messages from being processed multiple
times, the message type was changed to migrave_ros_msgs/StampedString,
which contains a message ID. The game base keeps track of the message
IDs and simply sends an acknowledgement if a repeated message is
received, but does not act on it again.